data_pipeline/requirements.txt
aiosqlite==0.20.0
alembic==1.13.1
anyio==4.3.0
apprise==1.7.4
asgi-lifespan==2.1.0
astunparse==1.6.3
async-timeout==4.0.3
asyncpg==0.29.0
backports.zoneinfo==0.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coolname==2.2.0
croniter==2.0.3
cryptography==42.0.5
dateparser==1.2.0
db-dtypes==1.2.0
docker==7.0.0
exceptiongroup==1.2.0
fastapi==0.101.0
fsspec==2024.3.1
google-api-core==2.17.1
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.19.0
google-cloud-bigquery-storage==2.24.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
griffe==0.42.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.2
importlib-resources==6.3.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.4
oauthlib==3.2.2
orjson==3.9.15
packaging==24.0
pandas==1.5.2
pandas-gbq==0.18.1
pathspec==0.12.1
pendulum==3.0.0
prefect==2.7.7
prefect-sqlalchemy==0.2.2
proto-plus==1.23.0
protobuf==4.21.11
psycopg2-binary==2.9.5
pyarrow==10.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
pydata-google-auth==1.8.2
pygments==2.17.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
readchar==4.0.6
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
rich==13.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.46
starlette==0.27.0
text-unidecode==1.3
time-machine==2.14.0
toml==0.10.2
typer==0.9.0
typing-extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.29.0
websocket-client==1.7.0
zipp==3.18.1
